+++ familyhelper-finance-impl/src/main/java/com/dwarfeng/familyhelper/finance/impl/util/FtpConstants.java
@@ -1,0 +1,16 @@
+package com.dwarfeng.familyhelper.finance.impl.util;
+
+/**
+ * FTP 常量。
+ *
+ * @author DwArFeng
+ * @since 1.4.0
+ */
+public final class FtpConstants {
+
+    public static final String PATH_BILL_FILE = "familyhelper-finance/bill-file";
+
+    private FtpConstants() {
+        throw new IllegalStateException("禁止实例化");
+    }
+}