+++ familyhelper-finance-impl/src/test/java/com/dwarfeng/familyhelper/finance/impl/configuration/ServiceExceptionMapperConfiguration.java
@@ -24,6 +24,7 @@ public class ServiceExceptionMapperConfiguration {
         destination.put(FundChangeNotExistsException.class, ServiceExceptionCodes.FUND_CHANGE_NOT_EXISTS);
         destination.put(IllegalFundChangeStateException.class, ServiceExceptionCodes.ILLEGAL_FUND_CHANGE_STATE);
         destination.put(InvalidPermissionLevelException.class, ServiceExceptionCodes.INVALID_PERMISSION_LEVEL);
+        destination.put(BillFileNotExistsException.class, ServiceExceptionCodes.BILL_FILE_NOT_EXISTS);
         return new MapServiceExceptionMapper(destination, com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.UNDEFINE);
     }
 }